python-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
deepface>=0.0.79
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
pydantic>=2.0.0
pillow>=10.0.0
tf-keras>=2.20.1
